+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 - [MINOR] Added `CollapsibleBox`
 - [MINOR] Added `SelectableView`
+- [MINOR] Added `Baseline` to `Alignment`
 
 ### Changed
 - [MINOR] Fixed `Markdown` to render images at maxWidth=100%
@@ -25,6 +26,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MAJOR] Updated `buildTheme` to ignore extra components, they should be added after calling buildTheme
 - [MAJOR] Completely changed how css is rendered - see new documentation for detail
 - [MAJOR] Change all displayName for component `X` to `KibaX`
+- [MINOR] Added `childAlignment` prop to `Button`
+- [MINOR] Updated `Icon` to size svg directly instead of container
 
 ### Removed